backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.20
openpyxl==3.1.5
